Right Strip Characters from a String

strs_rstrip removes trailing characters (spaces by default) from each element of a character vector. It is similar to Python's str.rstrip() method.

Usage

strs_rstrip(string, chars = NULL)

Arguments

string

A character vector where each element is a string to be right-stripped.

chars

An optional string of characters to be removed from the end of each element. If NULL, whitespace is removed.

Value

A character vector of the same length as string, with specified characters removed from the end of each element.

See also

Examples

strs_rstrip("hello world    ")
#> [1] "hello world"
strs_rstrip("hello worldxxx", chars = "x")
#> [1] "hello world"